requirements.txt
fastmcp>=2.3.0
openai>=1.0
python-dotenv>=1.0
rich>=13.0
streamlit>=1.28
hypothesis>=6.0
pytest>=7.0
